requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
langchain
langchain-google-genai
playwright
